+++ src/features/giftcards/GiftcardAdmin/index.ts
@@ -1,0 +1,1 @@
+export { GiftcardAdmin } from "./GiftcardAdmin";